+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -311,6 +311,35 @@ const DARK_BLUE_THEME_CONFIG = {
       'rgba(250, 251, 252, 0.5)',
   },
   delegationSetupWizard: {
+    '--theme-delegation-steps-choose-stake-pool-checkmark-icon-color':
+      '#536370',
+    '--theme-delegation-steps-choose-stake-pool-delegated-pools-label-color':
+      'rgba(233, 244, 254, 1)',
+    '--theme-delegation-steps-choose-stake-pool-select-box-placeholder-color':
+      '#9eabbb',
+    '--theme-delegation-steps-choose-stake-pool-selected-checkmark-icon-color':
+      '#fafbfc',
+    '--theme-delegation-steps-choose-stake-pool-selected-slug-color': '#fafbfc',
+    '--theme-delegation-steps-choose-stake-pool-slug-color': '#536370',
+    '--theme-delegation-steps-choose-stake-pool-title-color':
+      'rgba(233, 244, 254, 0.8)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-arrow-color':
+      'rgba(94, 96, 102, 0.9)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-background-color':
+      'rgba(94, 96, 102, 0.9)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-box-shadow':
+      '0 5px 20px 0 rgba(0, 0, 0, 0.25)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-description-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-slug-color':
+      'rgba(250, 251, 252, 0.6)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-table-label-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-table-value-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-title-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-url-color': '#2ab467',
     '--theme-delegation-steps-choose-wallet-custom-value-color': '#e9f4fe',
     '--theme-delegation-steps-choose-wallet-description-color':
       'rgba(233, 244, 254, 0.8)',
@@ -485,7 +514,7 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-node-update-button-text-color': '#fafbfc',
   },
   notification: {
-    '--theme-notification-message-background-color': 'rgba(83, 99, 112, 0.88)',
+    '--theme-notification-message-background-color': 'rgba(83, 99, 112, 0.95)',
     '--theme-notification-message-text-color': '#fafbfc',
     '--theme-legacy-badge-background-color': '#ab1700',
     '--theme-legacy-notification-background-color': '#ab2712',
@@ -572,35 +601,39 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-sidebar-menu-add-button-text-color': '#e9f4fe',
   },
   stakePools: {
-    '--theme-staking-stake-pools-title-color': '#e9f4fe',
-    '--theme-staking-stake-pools-search-button-color': '#e9f4fe',
     '--theme-staking-stake-pool-background-color': '#263345',
-    '--theme-staking-stake-pool-border-color': 'transparent',
+    '--theme-staking-stake-pool-border-color': '#536370',
     '--theme-staking-stake-pool-glow-color': '#7cfeb54c',
-    '--theme-staking-stake-pool-slug-color': '#e9f4fe',
     '--theme-staking-stake-pool-retirement-background-color': '#ea4c5b',
+    '--theme-staking-stake-pool-selected-background-color': '#5da377',
+    '--theme-staking-stake-pool-selected-checkmark-icon-color': '#fafbfc',
+    '--theme-staking-stake-pool-selected-slug-color': '#fafbfc',
+    '--theme-staking-stake-pool-slug-color': '#e9f4fe',
     '--theme-staking-stake-pool-tooltip-background-color':
       'rgba(83, 99, 112, 0.98)',
     '--theme-staking-stake-pool-tooltip-border-color': 'transparent',
-    '--theme-staking-stake-pool-tooltip-shadow-color': 'rgba(0, 0, 0, 0.18)',
-    '--theme-staking-stake-pool-tooltip-text-color': '#e9f4fe',
-    '--theme-staking-stake-pool-tooltip-link-color': '#2ab467',
-    '--theme-staking-stake-pool-tooltip-table-title-color': '#e9f4fe',
-    '--theme-staking-stake-pool-tooltip-table-param-color': '#e9f4fe',
-    '--theme-staking-stake-pool-tooltip-retirement-text-color': '#fafbfc',
-    '--theme-staking-stake-pool-tooltip-retirement-background-color': '#ea4c5b',
-    '--theme-staking-stake-pool-tooltip-delegate-button-background-color':
-      'rgba(0, 0, 0, 0.1)',
-    '--theme-staking-stake-pool-tooltip-delegate-button-hover-background-color':
-      '#fafbfc',
     '--theme-staking-stake-pool-tooltip-delegate-button-active-background-color': `${chroma(
       '#fafbfc'
     ).alpha(0.8)}`,
-    '--theme-staking-stake-pool-tooltip-delegate-button-text-color': '#e9f4fe',
-    '--theme-staking-stake-pool-tooltip-delegate-button-inverse-text-color':
-      '#263345',
+    '--theme-staking-stake-pool-tooltip-delegate-button-background-color':
+      'rgba(0, 0, 0, 0.1)',
     '--theme-staking-stake-pool-tooltip-delegate-button-border-color':
       '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-delegate-button-hover-background-color':
+      '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-delegate-button-inverse-text-color':
+      '#263345',
+    '--theme-staking-stake-pool-tooltip-delegate-button-text-color': '#e9f4fe',
+    '--theme-staking-stake-pool-tooltip-link-color': '#2ab467',
+    '--theme-staking-stake-pool-tooltip-retirement-background-color': '#ea4c5b',
+    '--theme-staking-stake-pool-tooltip-retirement-text-color': '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-shadow-color': 'rgba(0, 0, 0, 0.18)',
+    '--theme-staking-stake-pool-tooltip-table-param-color': '#e9f4fe',
+    '--theme-staking-stake-pool-tooltip-table-title-color': '#e9f4fe',
+    '--theme-staking-stake-pool-tooltip-text-color': '#e9f4fe',
+    '--theme-staking-stake-pools-search-button-color': '#e9f4fe',
+    '--theme-staking-stake-pools-search-icon-color': '#8793a1',
+    '--theme-staking-stake-pools-title-color': '#e9f4fe',
   },
   staking: {
     '--theme-staking-background-color': 'rgba(233, 244, 254, 0.1)',